requirements.txt
numpy==1.15.4
pycairo==1.18.0
